worked on the update todo status today

+++ src/components/card/Card.tsx
@@ -3,23 +3,38 @@ import { StyledCard, StyledCardHeader, StyledCardContent } from './styles';
 import { ITodos } from '../../types'
 import { FaTimesCircle, FaCheck, FaTrash } from 'react-icons/fa';
 import { useDispatch } from 'react-redux';
-import { deleteTodo } from '../../feature/action/todoAction';
+import { deleteTodo, updateTodo } from '../../feature/action/todoAction';
 import { truncate } from '../../utils/truncate';
 import moment from 'moment';
 
 const Card: FC<{ todo: ITodos }> = ({ todo }) => {
     const dispatch = useDispatch();
     const handleDeleteTodo = (id: string) => {
-        // @ts-ignore
-        dispatch(deleteTodo(id));
+
+        if(window.confirm('Are you sure you want to delete this Todo?')) {
+            // @ts-ignore
+            dispatch(deleteTodo(id));
+        } else {
+            return false;
+        }
     };
+
+    const handleCompleteTodo = (isComplete: boolean | undefined, id: string) => {
+        // @ts-ignore
+        dispatch(updateTodo(isComplete, id))
+    }
     return (
         <StyledCard>
             <StyledCardHeader>
                 <h4>{todo.title}</h4>
                 <span>
                     <FaTrash size={15} onClick={() => handleDeleteTodo(todo._id)} />{' '}
-                    {todo.isComplete ? <FaTimesCircle size={15} /> : <FaCheck size={15} />}
+                    {
+                        todo.isComplete ?
+                        <FaTimesCircle size={15} onClick={() => handleCompleteTodo(todo.isComplete, todo._id)} />
+                        :
+                        <FaCheck onClick={() => handleCompleteTodo(todo.isComplete, todo._id)} size={15} />
+                    }
                 </span>
             </StyledCardHeader>
             <StyledCardContent>
@@ -27,6 +42,7 @@ const Card: FC<{ todo: ITodos }> = ({ todo }) => {
                 <span>
                     <p>{moment(todo.createdAt).format('MMMM Do YYYY, h:mm:ss a')}</p>
                 </span>
+                <span>{todo.isComplete ? 'Completed' : 'Not Completed'}</span>
             </StyledCardContent>
         </StyledCard>
     );
